agrofresh-master/requirements.txt
django
django-admin-sortable2
django-crispy-forms
django-debug-toolbar
psycopg2
pymodbus
python-decouple
python-dateutil
pandas
plotly
uvicorn
websockets
whitenoise
rx